# Make /source about 3x faster (by making visible less accurate)