--- a/WebDriverAgentLib/Categories/XCUIElement+FBIsVisible.h
+++ b/WebDriverAgentLib/Categories/XCUIElement+FBIsVisible.h
@@ -25,6 +25,9 @@ NS_ASSUME_NONNULL_BEGIN
 /*! Whether or not the element is visible */
 @property (atomic, readonly) BOOL fb_isVisible;
 
+/*! Whether or not the element is visible, optionally using a 
+ *  less precise heuristic o bypass a very expensive hittest */
+- (BOOL)fb_isVisible:(BOOL)useHeuristic;
 @end
 
 NS_ASSUME_NONNULL_END