--- a/WebDriverAgentLib/Categories/XCUIElement+FBWebDriverAttributes.m
+++ b/WebDriverAgentLib/Categories/XCUIElement+FBWebDriverAttributes.m
@@ -100,6 +100,11 @@
   return self.fb_isVisible;
 }
 
+- (BOOL)isWDVisible:(BOOL)useHeuristic
+{
+  return [self fb_isVisible:useHeuristic];
+}
+
 - (BOOL)isWDAccessible
 {
   // Special cases: