The `/source` route is very slow; on older devices (iPhone 5S), I get response times of approx 3.7s for retrieving the source of a home page; on newer devices (iPhone 7), the response times are about 1s but still slow.

Most of the time is spent in:
- Getting the active application
- Determining the visibility of the individual elements

I didn't dig into fetching the application yet, but when determining the visibility, the visible rectangle and hitpoint are determined for each element with a non-zero size and this is a *very* expensive operation, especially in a tight loop.

This PR adds a "heuristic" flag to isVisible; if this flag is set:
- `visibleFrame` is assumed to be equal to `frame`
- the hit point is always assumed to be inside the application frame.

This flag is *only* used inside the `/source` route, not in other routes.

The reasoning is that when you want to get the `/source`, you are usually a bit less strict about the accuracy of the visibility and can always call isVisible for elements about which you care. Other options exist as well - we could make it a configuration flag, add a toggle to the route or create a newer route.

Here are the response times for the `/source` route before and after, on iOS 10:

| Device         | Before                  | After
|------------|-------------------|-------
| iPhone 5S  | 3770ms                  | 850ms
| iPhone 7     | 1000ms                  | 350ms

In the after time, most time is spend getting the active application, we still may be able to optimize that further.