@qmfrederik 2017-05-18
Again - putting this PR out there initially to gather feedback, let me know what your thoughts are and if/how we can land this.