@mykola-mokhnach 2017-05-20
The change means the actual isDisplayed property of an element and the corresponding property in the /source output might be different. Is this really what we want?