@qmfrederik 2017-05-20
@mykola-mokhnach Not really, but the performance of the `/source` route is a problem for us and something we'd like to address.

Other options include
1 Removing `isVisible` from `/source` alltogether
2 Use something else, like `isHidden` in `source` to indicate whether an element is explicitly marked as hidden (and exclude the hitpoint test); at least one WDA variant does it that way
3 Using stuff like `/source?heuristic=true` to indicate whether you want this behavior or not

(2) may be the better option here, what do you think?

... and others.

I'm wondering, on the Appium side of things, do you get feedback on the `/source` performance? Is it something you consider worth improving?