@mykola-mokhnach 2017-05-21
> I'm wondering, on the Appium side of things, do you get feedback on the /source performance? Is it something you consider worth improving?

Usually /source call is used for debug purposes and "precision" there is more important than performance. 

Have you also tried to measure method performance if **shouldUseTestManagerForVisibilityDetection** property is set to YES?