@qmfrederik 2017-05-21
> Have you also tried to measure method performance if shouldUseTestManagerForVisibilityDetection property is set to YES?

Yes, if this value is set to YES the performance is similar or worse, so > 4 sec on iPhone 5S and ~ 1 sec on iPhone 7