@marekcirkos 2017-05-22
I am sorry, but serving bad information is a bad idea. 
We should either focus in improving it's performance in some other way, but keeping correctness
or add param to `/source` that will not return visibility param (if you really really really want this).