@marekcirkos 2017-05-22
However it would be much better trying to understand why it is slow, and trying to resolve it. It will improve over whole WDA performance.