@qmfrederik 2017-05-22
@marekcirkos Sure, I understand the concerns about `visible` not being accurate, so let's try to figure out what else we can do. #597 already looks like a good start (I'll try to get some perf data on it).

I measured the "slow" parts in the code by basically calculating how long each block of code takes in code itself; I did not manage to get a performance profiler to attach to WDA.

I tried running a performance trace on WDA but as soon as you launch Instruments in Xcode, the XCUITest session is interrupted - did you manage to get instruments to do performance profiling of WDA tests?