@marekcirkos 2017-05-30
Closing as per discussion.

I don't think you can profile WDA with Xcode as it is a test. 
You can measure performance by calculating time spent in the code with `NSTimeIntervals`